pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apex-backend"
version = "1.0.0"
description = "APEX — AI Personal Assistant OS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Supabase
    "supabase>=2.10.0",
    "PyJWT[crypto]>=2.10.0",
    "cryptography>=43.0.0",
    # AI
    "anthropic>=0.40.0",
    "voyageai>=0.3.0",
    # Cache
    "redis[hiredis]>=5.2.0",
    # HTTP
    "httpx>=0.28.0",
    # Scheduling
    "apscheduler>=3.10.4",
    # Logging
    "structlog>=24.4.0",
    # Utilities
    "tenacity>=9.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
